Add files via upload
Code changed as far as letting it speak back to the user. This is where we hit the limitation of the Raspi as the PyObjC module can't be installed in this OS.

additional modules that got added/imported are:

playsound
gTTS
time( again, to make the program reset within 1sec)


the print statements have been changed with the function "lulu_speak", which is currently not working due to the missing module PyObjC. Further tests will be done with Windows as the OS but as of now, the project on the RASPI is being limited to non speech.